fix(store): bind prepared workflow snapshots to their identity
