+++ crates/horizon-core/src/cloud_run/store.rs
@@ -2,8 +2,10 @@
 
 mod database;
 mod remote_workspaces;
+mod workflow_writes;
 
 pub use remote_workspaces::{RemoteWorkspaceStoreError, StoredRemoteWorkspace};
+use workflow_writes::PreparedWorkflowInsert;
 
 use std::collections::HashMap;
 use std::path::{Path, PathBuf};
@@ -90,11 +92,11 @@ impl CloudWorkflowStore {
     /// Rejects invalid snapshots, duplicate workflow identities, and storage
     /// failures.
     pub fn create(&self, workflow: &CloudWorkflow) -> Result<StoredWorkflow, CloudStoreError> {
-        let snapshot = encode_workflow(workflow)?;
+        let prepared = PreparedWorkflowInsert::new(workflow)?;
         let mut connection = self.connection()?;
         let transaction = connection.transaction_with_behavior(TransactionBehavior::Immediate)?;
         ensure_current_schema(&transaction)?;
-        let stored = insert_workflow(&transaction, workflow, &snapshot)?;
+        let stored = prepared.persist(&transaction)?;
         transaction.commit()?;
         Ok(stored)
     }
@@ -293,34 +295,6 @@ struct WorkflowRow {
     updated_at_millis: i64,
     retain_until_millis: i64,
     snapshot: Vec<u8>,
-}
-
-// The caller must validate the schema before composing writes in this transaction.
-fn insert_workflow(
-    transaction: &rusqlite::Transaction<'_>,
-    workflow: &CloudWorkflow,
-    snapshot: &[u8],
-) -> Result<StoredWorkflow, CloudStoreError> {
-    let id = workflow.id.to_string();
-    if workflow_row(transaction, &id)?.is_some() {
-        return Err(CloudStoreError::WorkflowExists(workflow.id));
-    }
-    transaction.execute(
-        "INSERT INTO cloud_workflows (
-            workflow_id, revision, created_at_millis, updated_at_millis, retain_until_millis, snapshot
-         ) VALUES (?1, 1, ?2, ?3, ?4, ?5)",
-        params![
-            id,
-            workflow.created_at_millis,
-            workflow.updated_at_millis,
-            workflow.retain_until_millis,
-            snapshot
-        ],
-    )?;
-    Ok(StoredWorkflow {
-        workflow: workflow.clone(),
-        revision: 1,
-    })
 }
 
 impl WorkflowRow {
